# swagger.json
{
	"swagger": "2.0",
	"info": {
		"version": "1.0.0",
		"title": "Makata Calendar API"
	},
	"schemes": ["https"],
	"host": "https://lab17-makata.herokuapp.com/",
	"tags": [
		{
			"name": "Authentication",
			"description": "User registration and log in"
		},
		{
			"name": "Calendar",
			"description": "Handling events"
		},
		{
			"name": "Events",
			"description": "Handling events"
		},
		{
			"name": "Subscribe",
			"description": "Subscribe or unsubscribe another user's calendar"
		}
	],
	"paths": {
		"/auth/register": {
			"post": {
				"tags": ["Authentication"],
				"summary": "Register a new user",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"parameters": [
					{
						"in": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/User"
						}
					}
				],
				"responses": {
					"200": {
						"description": "successfully created a user",
						"schema": {
							"$ref": "#/definitions/AuthenticatedUser"
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"tags": ["Authentication"],
				"summary": "User Login",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"parameters": [
					{
						"in": "body",
						"required": true,
						"description": "UserId can be a registered username or a registered email",
						"schema": {
							"type": "object",
							"properties": {
								"userId": {
									"type": "string"
								},
								"password": {
									"type": "string"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "User is successfully login",
						"schema": {
							"$ref": "#/definitions/AuthenticatedUser"
						}
					}
				}
			}
		}
	},
	"definitions": {
		"User": {
			"type": "object",
			"properties": {
				"firstName": {
					"type": "string"
				},
				"lastName": {
					"type": "string"
				},
				"email": {
					"type": "string"
				},
				"username": {
					"type": "string"
				},
				"password": {
					"type": "string"
				}
			}
		},
		"AuthenticatedUser": {
			"type": "object",
			"properties": {
				"profile": {
					"type": "object",
					"properties": {
						"firstName": {
							"type": "string"
						},
						"lastName": {
							"type": "string"
						},
						"email": {
							"type": "string"
						}
					}
				},
				"accessToken": {
					"type": "string"
				}
			}
		}
	}
}